api_yamdb/requirements.txt
requests==2.26.0
Django==2.2.16
djangorestframework==3.12.4
PyJWT==2.1.0
django-filter==2.4.0
djangorestframework-simplejwt==5.0.0
python-dotenv==0.19.2
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
pytest-django==4.4.0
pytest==6.2.4
pytest-pythonpath==0.7.3
